requirements.txt
PyYAML~=6.0.1
pymongo~=4.8.0
loguru~=0.7.2
fastapi~=0.112.0
uvicorn[standard]~=0.30.6
pydantic~=2.8.2
pydantic[email]
starlette~=0.37.2
SQLAlchemy~=2.0.32
PyJWT~=2.9.0
BAC0~=23.7.3
pytz
mariadb
paho-mqtt~=2.1.0
pyserial~=3.5
pyMeterBus~=0.8.4
pymodbus~=3.7.3
pyinstaller~=6.11.0
